FM-2: add string resources and refactor architecture
