+++ app/src/main/res/values-en/strings.xml
@@ -5,4 +5,14 @@
     <string name="settings">Settings</string>
     <string name="history">History</string>
     <string name="about">About us</string>
+    <string name="neon_styles">Neon styles</string>
+    <string name="language">Language</string>
+    <string name="music">Music</string>
+    <string name="sound">Sound</string>
+    <string name="clear_progress">Clear progress</string>
+    <string name="report_problem">Report a problem</string>
+    <string name="how_to_play">How to play?</string>
+    <string name="rate_us">Rate us</string>
+    <string name="support">Support</string>
+    <string name="vibration">Vibration</string>
 </resources>
