+++ app/src/main/res/values/strings.xml
@@ -4,4 +4,18 @@
     <string name="settings">Ustawienia</string>
     <string name="history">Historia</string>
     <string name="about">O nas</string>
+    <string name="neon_styles">Neonowy styl przycisków</string>
+    <string name="language">Język</string>
+    <string name="vibration">Wibracje</string>
+    <string name="music">Muzyka</string>
+    <string name="sound">Dźwięk</string>
+    <string name="clear_progress">Wyczyść postęp</string>
+    <string name="report_problem">Zgłoś problem</string>
+    <string name="how_to_play">Jak grać?</string>
+    <string name="rate_us">Oceń nas</string>
+    <string name="support">Wspieraj</string>
+    <string-array name="languages" translatable="false">
+        <item>Polski</item>
+        <item>English</item>
+    </string-array>
 </resources>
